Deduplicate django hooks files
