+++ newrelic/hooks/framework_django.py
@@ -38,10 +38,6 @@ from newrelic.common.object_wrapper import (
 )
 from newrelic.config import extra_settings
 from newrelic.core.config import global_settings
-from newrelic.hooks.framework_django_py3 import (
-    _nr_wrap_converted_middleware_async_,
-    _nr_wrapper_BaseHandler_get_response_async_,
-)
 
 _logger = logging.getLogger(__name__)
 
@@ -418,6 +414,36 @@ def _nr_wrapper_BaseHandler_get_response_(wrapped, instance, args, kwargs):
         delattr(request, "_nr_exc_info")
 
     return response
+
+
+async def _nr_wrapper_BaseHandler_get_response_async_(wrapped, instance, args, kwargs):
+    response = await wrapped(*args, **kwargs)
+
+    if current_transaction() is None:
+        return response
+
+    request = _bind_get_response(*args, **kwargs)
+
+    if hasattr(request, "_nr_exc_info"):
+        notice_error(error=request._nr_exc_info, status_code=response.status_code)
+        delattr(request, "_nr_exc_info")
+
+    return response
+
+
+def _nr_wrap_converted_middleware_async_(middleware, name):
+    @function_wrapper
+    async def _wrapper(wrapped, instance, args, kwargs):
+        transaction = current_transaction()
+
+        if transaction is None:
+            return await wrapped(*args, **kwargs)
+
+        transaction.set_transaction_name(name, priority=2)
+
+        return await FunctionTraceWrapper(wrapped, name=name)(*args, **kwargs)
+
+    return _wrapper(middleware)
 
 
 # Post import hooks for modules.
